refactor(view): 입력 받은 값에 대한 예외 처리 추가
- InputView 클래스: 사용자가 입력한 횟수가 범위를 벗어났거나, 숫자가 아닌 경우 예외처리를 추가함.
